pip_requirements.txt
appdirs==1.4.4
beautifulsoup4==4.9.1
black==20.8b1
click==7.1.2
flake8==3.8.3
html2text==2020.1.16
html5lib==1.1
importlib-metadata==1.7.0
lxml==4.5.2
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.8.0
pycodestyle==2.6.0
pyflakes==2.2.0
PyYAML==5.3.1
regex==2020.7.14
six==1.15.0
soupsieve==2.0.1
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
webencodings==0.5.1
zipp==3.1.0
